refactor(xlsx): extract conditional-formatting engine from renderer.ts
Move CF rule compilation and per-cell evaluation (~345 lines) into
src/conditional-format.ts: cellIs/colorScale/dataBar/iconSet handling,
cfvo resolution, dxf application. Exposes CfContext/CfResult/
CompiledCfRule, compileCf, evaluateCf; imports evalFormulaToBool from
formula.js.

renderer.ts drops from 3603 to 3258 lines. Typecheck passes, VRT
regression 65/65 at 0.0% diff.
